refactor : 토큰 생성을 하나로 처리
