+++ src/main/java/com/example/demo/domain/token/domain/dto/TokenResponse.java
@@ -1,0 +1,10 @@
+package com.example.demo.domain.token.domain.dto;
+
+public record TokenResponse(
+        String accessToken,
+        String refreshToken
+) {
+    public static TokenResponse create(String accessToken, String refreshToken) {
+        return new TokenResponse(accessToken, refreshToken);
+    }
+}
