--- src/main/java/com/example/demo/global/jwt/dto/TokenResponse.java
@@ -1,7 +1,0 @@
-package com.example.demo.global.jwt.dto;
-
-public record TokenResponse(
-        String accessToken,
-        String refreshToken
-) {
-}
